Create notification manager and sender.
The notification manager module will handle newly received meeting
(event) details and locations. It store this data in a meetingstore or
locationstore respectively, and schedule checks about participant's
locations leading up to a meeting. If from one of these checks it is
determined that a notification needs to be sent, the notificationsender
will handle the formatting and sending of this email.